pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pde-solver"
version = "0.1.0"
description = "Research-grade hybrid physics + AI PDE solver for Navier-Stokes, GR, quantum equations"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "PDE Solver Contributors"}
]
keywords = ["pde", "neural-networks", "physics-informed", "scientific-computing", "cfd"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "torch>=2.0.0",
    "sympy>=1.12",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "omegaconf>=2.3.0",
    "typer>=0.9.0",
    "wandb>=0.15.0",
    "tqdm>=4.65.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "prometheus-client>=0.17.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.23.0",
]

[project.optional-dependencies]
jax = ["jax>=0.4.0", "jaxlib>=0.4.0"]
firedrake = ["firedrake>=2023.0.0"]
mpi = ["mpi4py>=3.1.0"]
dev = [
    "pre-commit>=3.3.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
]
plotly = ["plotly>=5.17.0"]

[project.scripts]
pde-solver = "pde_solver.cli:app"

[tool.setuptools]
packages = ["pde_solver"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=pde_solver --cov-report=term-missing --cov-report=html"


